moved squares inside pkg
