+++ internal/pkg/squares/squares.go
@@ -29,6 +29,8 @@ func New(size int) Squareser {
 	return new(size)
 }
 
+// binChop calling sort.Search is slower than hand built binChop
+// by 4.8021043 seconds against 2.6710666 seconds for a size of 10000
 func (sq Squares) binChop(value int64, start int, end int) (int64, bool, int) {
 	set := sq.sqs[start:end]
 	offset := start
