+++ internal/pkg/squares/squares_test.go
